Add API methods for livechat users (agents or managers) management
the delete method needed to be introduced in the base API as this time the deletions are done the REST way. Also because of this, the signature of the methods has changed to have a verb in it as the endpoint keeps being the same, only the HTTP method changes.